requirements.txt
streamlit
pg8000
requests
python-dotenv
